add cast CLogicInt

+++ sakura_core/doc/layout/CLayout.cpp
@@ -25,8 +25,8 @@ CLayout::~CLayout()
 void CLayout::DUMP( )
 {
 	DEBUG_TRACE( L"\n\n■CLayout::DUMP()======================\n" );
-	DEBUG_TRACE( L"m_ptLogicPos.y=%d\t\t対応する論理行番号\n", m_ptLogicPos.y );
-	DEBUG_TRACE( L"m_ptLogicPos.x=%d\t\t対応する論理行の先頭からのオフセット\n", m_ptLogicPos.x );
+	DEBUG_TRACE( L"m_ptLogicPos.y=%d\t\t対応する論理行番号\n", static_cast<int>(m_ptLogicPos.y) );
+	DEBUG_TRACE( L"m_ptLogicPos.x=%d\t\t対応する論理行の先頭からのオフセット\n", static_cast<int>(m_ptLogicPos.x) );
 	DEBUG_TRACE( L"m_nLength=%d\t\t対応する論理行のハイト数\n", (int)m_nLength );
 	DEBUG_TRACE( L"m_nTypePrev=%d\t\tタイプ 0=通常 1=行コメント 2=ブロックコメント 3=シングルクォーテーション文字列 4=ダブルクォーテーション文字列 \n", m_nTypePrev );
 	DEBUG_TRACE( L"======================\n" );
